Refactored bigquery client Solves #1073
Major changes:
  - Aligned biquery job types under the same trait
  - Grouped methods by functionality (query, load, etc)
  - Moved cache and config logic outside of bqclient
