+++ scio-bigquery/src/it/scala/com/spotify/scio/bigquery/types/BigQueryTypeIT.scala
@@ -107,7 +107,7 @@ class BigQueryTypeIT extends FlatSpec with Matchers {
 
   it should "round trip rows with legacy syntax" in {
     val bqt = BigQueryType[LegacyT]
-    val rows = bq.getQueryRows(legacyQuery).toList
+    val rows = bq.query.getRows(legacyQuery).toList
     val typed = Seq(LegacyT("Romeo", 117L))
     rows.map(bqt.fromTableRow) shouldBe typed
     typed.map(bqt.toTableRow).map(bqt.fromTableRow) shouldBe typed
@@ -115,7 +115,7 @@ class BigQueryTypeIT extends FlatSpec with Matchers {
 
   it should "round trip rows with SQL syntax" in {
     val bqt = BigQueryType[SqlT]
-    val rows = bq.getQueryRows(sqlQuery).toList
+    val rows = bq.query.getRows(sqlQuery).toList
     val typed = Seq(SqlT(Some("Romeo"), Some(117L)))
     rows.map(bqt.fromTableRow) shouldBe typed
     typed.map(bqt.toTableRow).map(bqt.fromTableRow) shouldBe typed
