+++ scio-bigquery/src/main/scala/com/spotify/scio/bigquery/MockBigQuery.scala
@@ -61,8 +61,8 @@ class MockBigQuery private (private val bq: BigQueryClient) {
       !mapping.contains(original),
       s"Table ${BigQueryHelpers.toTableSpec(original)} already registered for mocking")
 
-    val t = bq.getTable(original)
-    val temp = bq.temporaryTable(t.getLocation)
+    val t = bq.tables.get(original)
+    val temp = bq.tables.createTemporary(t.getLocation)
     mapping += (original -> temp)
     new MockTable(bq, t.getSchema, original, temp)
   }
@@ -72,12 +72,12 @@ class MockBigQuery private (private val bq: BigQueryClient) {
    * data.
    */
   def queryResult(sqlQuery: String, flattenResults: Boolean = false): Seq[TableRow] = {
-    val isLegacy = bq.isLegacySql(sqlQuery, flattenResults)
+    val isLegacy = bq.query.isLegacySql(sqlQuery, flattenResults)
     val mockQuery = mapping.foldLeft(sqlQuery){ case (q, (src, dst)) =>
       q.replace(toTableSpec(src, isLegacy), toTableSpec(dst, isLegacy))
     }
     try {
-      bq.getQueryRows(mockQuery, flattenResults).toList
+      bq.query.getRows(mockQuery, flattenResults).toList
     } catch {
       case e: GoogleJsonResponseException if new ApiErrorExtractor().itemNotFound(e) =>
         throw new RuntimeException(
@@ -122,7 +122,7 @@ class MockTable(private val bq: BigQueryClient,
   }
 
   private def writeRows(rows: Seq[TableRow]): Unit = {
-    bq.writeTableRows(temp, rows.toList, schema, WRITE_EMPTY, CREATE_IF_NEEDED)
+    bq.tables.writeRows(temp, rows.toList, schema, WRITE_EMPTY, CREATE_IF_NEEDED)
   }
 
   /**
@@ -147,7 +147,7 @@ class MockTable(private val bq: BigQueryClient,
    */
   def withSample(numRows: Int): Unit = {
     ensureUnique()
-    val rows = bq.getTableRows(original).take(numRows).toList
+    val rows = bq.tables.getRows(original).take(numRows).toList
     require(rows.length == numRows, s"Sample size ${rows.length} != requested $numRows")
     writeRows(rows)
   }
@@ -158,7 +158,7 @@ class MockTable(private val bq: BigQueryClient,
    */
   def withSample(minNumRows: Int, maxNumRows: Int): Unit = {
     ensureUnique()
-    val rows = bq.getTableRows(original).take(maxNumRows).toList
+    val rows = bq.tables.getRows(original).take(maxNumRows).toList
     require(
       rows.length >= minNumRows && rows.length <= maxNumRows,
       s"Sample size ${rows.length} < requested minimal $minNumRows")
