+++ scio-bigquery/src/main/scala/com/spotify/scio/bigquery/package.scala
@@ -36,8 +36,8 @@ import scala.util.Try
  * [[https://cloud.google.com/bigquery/docs/reference/legacy-sql legacy]] and
  * [[https://cloud.google.com/bigquery/docs/reference/standard-sql/ standard]].
  * APIs that take a BigQuery query string as argument, e.g.
- * [[com.spotify.scio.bigquery.BigQueryClient.getQueryRows BigQueryClient.getQueryRows]],
- * [[com.spotify.scio.bigquery.BigQueryClient.getQuerySchema BigQueryClient.getQuerySchema]],
+ * [[com.spotify.scio.bigquery.BigQueryClient.query.getRows BigQueryClient.query.getRows]],
+ * [[com.spotify.scio.bigquery.BigQueryClient.query.getSchema BigQueryClient.query.getSchema]],
  * [[com.spotify.scio.bigquery.BigQueryClient.getTypedRows BigQueryClient.getTypedRows]] and
  * [[com.spotify.scio.bigquery.BigQueryType.fromQuery BigQueryType.fromQuery]], automatically
  * detects the query's dialect. To override this, start the query with either `#legacysql` or
