+++ scio-bigquery/src/main/scala/com/spotify/scio/bigquery/types/TypeProvider.scala
@@ -44,7 +44,7 @@ private[types] object TypeProvider {
 
     val args = extractStrings(c, "Missing table specification")
     val tableSpec = BigQueryPartitionUtil.latestTable(bigquery, formatString(args))
-    val schema = bigquery.getTableSchema(tableSpec)
+    val schema = bigquery.tables.getSchema(tableSpec)
     val traits = List(tq"${p(c, SType)}.HasTable")
     val overrides = List(q"override def table: _root_.java.lang.String = ${args.head}")
 
@@ -62,7 +62,7 @@ private[types] object TypeProvider {
 
     val args = extractStrings(c, "Missing query")
     val query = BigQueryPartitionUtil.latestQuery(bigquery, formatString(args))
-    val schema = bigquery.getQuerySchema(query)
+    val schema = bigquery.query.getSchema(query)
     val traits = Seq(tq"${p(c, SType)}.HasQuery")
     val overrides = Seq(q"override def query: _root_.java.lang.String = ${args.head}")
