+++ scio-test/src/it/scala/com/spotify/scio/testing/util/ItUtils.scala
@@ -25,7 +25,7 @@ import com.google.auth.Credentials
 import com.google.auth.http.HttpCredentialsAdapter
 import com.google.cloud.hadoop.util.ChainingHttpRequestInitializer
 import com.google.common.collect.ImmutableList
-import com.spotify.scio.bigquery.BigQueryClient
+import com.spotify.scio.bigquery.{BigQueryClient, BigQueryConfig}
 import org.apache.beam.sdk.extensions.gcp.auth.NullCredentialInitializer
 import org.apache.beam.sdk.extensions.gcp.options.GcpOptions.DefaultProjectFactory
 import org.apache.beam.sdk.extensions.gcp.options._
@@ -37,9 +37,9 @@ private[scio] object ItUtils {
 
   /** Get project for integration test. */
   def project: String =
-    if (sys.props(BigQueryClient.PROJECT_KEY) != null) {
+    if (sys.props(BigQueryConfig.PROJECT_KEY) != null) {
       // this is usually set by CI script
-      sys.props(BigQueryClient.PROJECT_KEY)
+      sys.props(BigQueryConfig.PROJECT_KEY)
     } else {
       // fallback to local setting
       new DefaultProjectFactory().create(null)
